requirements.txt
mistralai
dotenv
openai
groq
streamlit
streamlit_mic_recorder
matplotlib
django
django-rest-framework
django-cors-headers
djangorestframework-simplejwt